refactor(differential): extract shared ReportUI chrome
Move the differential report's product-neutral interactive chrome — CSS
design system, filter/toggle/theme JS, and the three-axis header — out of
the imgproxy-specific ReportHtml into a reusable
ImagePipe.Test.Differential.ReportUI so the TwicPics suite can adopt the
same report.

ReportUI owns css/0, script/0, and render/1 (title, provenance_html,
counts_html, a suite-supplied type/group axis, head_extras, and
pre-rendered cards in display order; status + heatmap axes are fixed).
The localStorage theme key is renamed imgproxy-report-theme →
differential-report-theme (no test pinned the old key).

imgproxy ReportHtml now builds its provenance/counts/type-axis vocab and
per-card HTML, sorts flagged-first, and delegates the shell to ReportUI;
its card/badges/triage/visuals/panel helpers stay put. Rendered output is
unchanged — the ReportHtml.render/1 unit tests and the :imgproxy_report
smoke test stay green.
